add shims to encoding
