+++ packages/crypto/package.json
@@ -46,9 +46,7 @@
     "@noble/curves": "1.9.0",
     "@noble/hashes": "1.8.0",
     "@turnkey/encoding": "workspace:*",
-    "borsh": "2.0.0",
-    "bs58check": "4.0.0",
-    "bs58": "6.0.0"
+    "borsh": "2.0.0"
   },
   "devDependencies": {
     "@turnkey/api-key-stamper": "workspace:*",
